AT2-39 refactor folder structure
