+++ ska-webapps-react-skeleton/src/components/App/App.test.jsx
